Move pipeline test util to its own project in order to properly test scope it and bfd-shared-test-utils
