+++ apps/bfd-pipeline/bfd-pipeline-shared-test-utils/pom.xml
@@ -1,0 +1,58 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<modelVersion>4.0.0</modelVersion>
+
+<parent>
+    <groupId>gov.cms.bfd</groupId>
+    <artifactId>bfd-pipeline-parent</artifactId>
+    <version>1.0.0-SNAPSHOT</version>
+</parent>
+
+<artifactId>bfd-pipeline-shared-test-utils</artifactId>
+<packaging>jar</packaging>
+
+<description>
+    Framework and utility test code shared across the BFD Pipeline application.
+</description>
+
+<dependencies>
+    <dependency>
+        <groupId>gov.cms.bfd</groupId>
+        <artifactId>bfd-model-rif</artifactId>
+        <version>${project.version}</version>
+    </dependency>
+    <dependency>
+        <groupId>gov.cms.bfd</groupId>
+        <artifactId>bfd-model-rda</artifactId>
+        <version>${project.version}</version>
+    </dependency>
+    <dependency>
+        <groupId>gov.cms.bfd</groupId>
+        <artifactId>bfd-pipeline-shared-utils</artifactId>
+        <version>${project.version}</version>
+    </dependency>
+    <dependency>
+        <groupId>gov.cms.bfd</groupId>
+        <artifactId>bfd-shared-test-utils</artifactId>
+        <version>${project.version}</version>
+    </dependency>
+
+    <dependency>
+        <!-- Used to provide DB connection pooling. See https://github.com/brettwooldridge/HikariCP
+                for details. -->
+        <groupId>com.zaxxer</groupId>
+        <artifactId>HikariCP</artifactId>
+        <version>${hikari.version}</version>
+    </dependency>
+
+    <dependency>
+        <!-- Provides a sane facade for the giant mess of incompatible logging
+            frameworks in Java. -->
+        <groupId>org.slf4j</groupId>
+        <artifactId>slf4j-api</artifactId>
+        <version>${slf4j.version}</version>
+    </dependency>
+
+</dependencies>
+
+</project>
