+++ apps/bfd-pipeline/pom.xml
@@ -20,6 +20,7 @@
 		<module>bfd-pipeline-rda-grpc</module>
 		<module>bfd-pipeline-rda-grpc-apps</module>
 		<module>bfd-pipeline-shared-utils</module>
+		<module>bfd-pipeline-shared-test-utils</module>
 		<module>bfd-pipeline-app</module>
 		<module>bfd-pipeline-rda-bridge</module>
 	</modules>
@@ -50,12 +51,6 @@
 				<groupId>gov.cms.bfd</groupId>
 				<artifactId>bfd-model-rif-samples</artifactId>
 				<version>${project.version}</version>
-			</dependency>
-			<dependency>
-				<groupId>gov.cms.bfd</groupId>
-				<artifactId>bfd-shared-test-utils</artifactId>
-				<version>${project.version}</version>
-				<!-- Scope cannot be test due to PipelineTestUtils, fix that to properly scope this -->
 			</dependency>
 		</dependencies>
 	</dependencyManagement>
