+++ apps/bfd-server/bfd-server-war/pom.xml
@@ -310,6 +310,12 @@
 			<version>${project.version}</version>
 			<!-- Scope cannot be test due to SpringConfiguration using test classes, fix that to properly scope this -->
 		</dependency>
+		<dependency>
+			<groupId>gov.cms.bfd</groupId>
+			<artifactId>bfd-pipeline-shared-test-utils</artifactId>
+			<version>${project.version}</version>
+			<scope>test</scope>
+		</dependency>
 	</dependencies>
 
 	<build>
